refactor: :building_construction: ディレクトリ構成を変更
