+++ lib/feature/flavor/infrastructure/flavor.dart
@@ -1,8 +1,8 @@
 import 'package:firebase_core/firebase_core.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../util/firebase_option/firebase_options_dev.dart' as dev_options;
-import '../../util/firebase_option/firebase_options_prod.dart' as prod_options;
+import '../../../util/firebase_option/firebase_options_dev.dart' as dev_options;
+import '../../../util/firebase_option/firebase_options_prod.dart' as prod_options;
 
 part 'flavor.g.dart';
